Register DiscoveryFile globally
