+++ src/meltano/core/yaml.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from ruamel.yaml import YAML
 
 from meltano.core.behavior.canonical import Canonical
+from meltano.core.discovery_file import DiscoveryFile
 from meltano.core.plugin import PluginType
 from meltano.core.setting_definition import SettingKind
 
@@ -21,5 +22,6 @@ def configure_yaml() -> YAML:
     yaml.register_class(Canonical)
     yaml.register_class(PluginType)
     yaml.register_class(SettingKind)
+    yaml.register_class(DiscoveryFile)
 
     return yaml
